rename everything that is downloadReferralsByAgency to simply downloadReferrals
