+++ src/containers/downloads/actions/index.js
@@ -5,12 +5,12 @@ import type { RequestSequence } from 'redux-reqseq';
 const DOWNLOAD_CASES :'DOWNLOAD_CASES' = 'DOWNLOAD_CASES';
 const downloadCases :RequestSequence = newRequestSequence(DOWNLOAD_CASES);
 
-const DOWNLOAD_REFERRALS_BY_AGENCY :'DOWNLOAD_REFERRALS_BY_AGENCY' = 'DOWNLOAD_REFERRALS_BY_AGENCY';
-const downloadReferralsByAgency :RequestSequence = newRequestSequence(DOWNLOAD_REFERRALS_BY_AGENCY);
+const DOWNLOAD_REFERRALS :'DOWNLOAD_REFERRALS' = 'DOWNLOAD_REFERRALS';
+const downloadReferrals :RequestSequence = newRequestSequence(DOWNLOAD_REFERRALS);
 
 export {
   DOWNLOAD_CASES,
-  DOWNLOAD_REFERRALS_BY_AGENCY,
+  DOWNLOAD_REFERRALS,
   downloadCases,
-  downloadReferralsByAgency,
+  downloadReferrals,
 };
