+++ src/containers/downloads/reducers/index.js
@@ -2,22 +2,22 @@
 import { Map, fromJS } from 'immutable';
 
 import downloadCasesReducer from './downloadCasesReducer';
-import downloadReferralsByAgencyReducer from './downloadReferralsByAgencyReducer';
+import downloadReferralsReducer from './downloadReferralsReducer';
 
 import { RESET_REQUEST_STATE } from '../../../core/redux/actions';
 import { RS_INITIAL_STATE } from '../../../core/redux/constants';
 import { resetRequestStateReducer } from '../../../core/redux/reducers';
 import {
   DOWNLOAD_CASES,
-  DOWNLOAD_REFERRALS_BY_AGENCY,
+  DOWNLOAD_REFERRALS,
   downloadCases,
-  downloadReferralsByAgency,
+  downloadReferrals,
 } from '../actions';
 
 const INITIAL_STATE :Map = fromJS({
   // actions
   [DOWNLOAD_CASES]: RS_INITIAL_STATE,
-  [DOWNLOAD_REFERRALS_BY_AGENCY]: RS_INITIAL_STATE,
+  [DOWNLOAD_REFERRALS]: RS_INITIAL_STATE,
   // data
 });
 
@@ -33,8 +33,8 @@ export default function reducer(state :Map = INITIAL_STATE, action :Object) {
       return downloadCasesReducer(state, action);
     }
 
-    case downloadReferralsByAgency.case(action.type): {
-      return downloadReferralsByAgencyReducer(state, action);
+    case downloadReferrals.case(action.type): {
+      return downloadReferralsReducer(state, action);
     }
 
     default:
